fix(ci-tests): one reading of the SQL a Go file sends, and it decodes escapes
`ast.BasicLit.Value` is the source text, not the string the database gets.
An interpreted literal keeps its escapes, so `"INSERT\nINTO
user_record_view …"` holds a backslash and an `n` where Postgres receives a
newline — and a pattern asking for `\s` finds neither. A second writer or a
second ownership probe written in double quotes passes a census untouched
while sending the very statement it exists to find.

Measured against the census that MERGED this morning: `INSERT\nINTO …`
matched=false, `INSERT INTO\nuser_record_view …` matched=false, the same
statements in backticks matched=true. The hole was already shipped.

Both censuses read through one function now. `statementsIn` and
`sqlLiteralsIn` were the same walk under two names in one package and had
already drifted — one decodes escapes and flattens `+` chains, the other did
neither. Two answers to "what does this file send the database" is the
defect a census exists to find, standing inside the censuses.

Verified both directions after the collapse: the escaped spellings now
match, and the near-neighbour `user_record_view_archive` still does not.
